converting audio format

+++ android/src/main/kotlin/com/isvisoft/flutter_screen_recording/FlutterScreenRecordingPlugin.kt
@@ -398,7 +398,7 @@ class FlutterScreenRecordingPlugin(
     @kotlin.Throws(IOException::class)
     fun writeString(output: DataOutputStream?, value: String) {
         for (i in 0 until value.length) {
-            output?.write(value[i])
+            output?.write(value[i].toInt())
         }
     }
